Move files to "stats" directory
