+++ listenbrainz/webserver/static/js/src/stats/Bar.tsx
